requirements.txt
asgiref==3.7.2
class-registry==2.1.2
Django==4.2.7
django-cors-headers==4.3.0
django-filter==23.3
djangorestframework==3.14.0
filters==1.3.2
localflavor==1.9
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
regex==2023.10.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
